+++ src/ServiceProvider/TransformerContainerServiceProvider.php
@@ -3,7 +3,7 @@
 namespace RCatlin\Blog\ServiceProvider;
 
 use League\Container\ServiceProvider\AbstractServiceProvider;
-use RCatlin\Blog\Serializer;
+use RCatlin\Blog\Transformer;
 
 class TransformerContainerServiceProvider extends AbstractServiceProvider
 {
@@ -11,7 +11,7 @@ class TransformerContainerServiceProvider extends AbstractServiceProvider
      * @var array
      */
     protected $provides = [
-        Serializer\TransformerContainer::class,
+        Transformer\TransformerContainer::class,
     ];
 
     /**
@@ -19,6 +19,6 @@ class TransformerContainerServiceProvider extends AbstractServiceProvider
      */
     public function register()
     {
-        $this->getContainer()->share(Serializer\TransformerContainer::class);
+        $this->getContainer()->share(Transformer\TransformerContainer::class);
     }
 }
