+++ test/Unit/Transformer/Entity/ArticleTransformerTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace RCatlin\Blog\Test\Unit\Serializer\Transformer\Entity;
+namespace RCatlin\Blog\Test\Unit\Transformer\Entity;
 
 use RCatlin\Blog\Entity;
-use RCatlin\Blog\Serializer\Transformer;
+use RCatlin\Blog\Transformer;
 use RCatlin\Blog\Test\Unit\HasFaker;
 
 class ArticleTransformerTest extends \PHPUnit_Framework_TestCase
